Write demo app.js
The hjs-webpack package includes a minimal index.html it will generate if we don’t describe a custom one (using the html key in the configuration object). We won’t need to generate a custom html file here, so instead we’ll use the basic generated page. The default template includes a <div /> component with an id of root just for the purpose of setting up a Single Page App (SPA, for short).